requirements.txt
django-ckeditor
django-cleanup
django-crispy-forms
django-js-asset